# frontend/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "JalNetra",
  "description": "Satellite-based water quality anomaly intelligence for Maharashtra.\n\n**Boundary:** the system detects *potential anomalies* in satellite-observable indicators and prioritises zones for ground investigation. It never claims to detect, prove or confirm pollution or contamination.",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/v1/water-bodies": {
   "get": {
    "tags": [
     "water-bodies"
    ],
    "summary": "List Water Bodies",
    "description": "Registered water bodies with tier, area, latest observation and status.",
    "operationId": "list_water_bodies_api_v1_water_bodies_get",
    "parameters": [
     {
      "name": "district",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "District"
      }
     },
     {
      "name": "tier",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer",
         "maximum": 3,
         "minimum": 1
        },
        {
         "type": "null"
        }
       ],
       "title": "Tier"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 200,
       "minimum": 1,
       "default": 50,
       "title": "Limit"
      }
     },
     {
      "name": "cursor",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Cursor"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/WaterBodyList"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/water-bodies/{water_body_id}": {
   "get": {
    "tags": [
     "water-bodies"
    ],
    "summary": "Get Water Body",
    "description": "Detail with the GeoJSON boundary and zone polygons.",
    "operationId": "get_water_body_api_v1_water_bodies__water_body_id__get",
    "parameters": [
     {
      "name": "water_body_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Water Body Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/WaterBodyDetail"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/water-bodies/{water_body_id}/observations": {
   "get": {
    "tags": [
     "water-bodies"
    ],
    "summary": "List Observations",
    "description": "Available dates with cloud cover, usability and the stage each scene reached.",
    "operationId": "list_observations_api_v1_water_bodies__water_body_id__observations_get",
    "parameters": [
     {
      "name": "water_body_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Water Body Id"
      }
     },
     {
      "name": "from",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "date"
        },
        {
         "type": "null"
        }
       ],
       "title": "From"
      }
     },
     {
      "name": "to",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "date"
        },
        {
         "type": "null"
        }
       ],
       "title": "To"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 500,
       "minimum": 1,
       "default": 100,
       "title": "Limit"
      }
     },
     {
      "name": "cursor",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Cursor"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ObservationList"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/water-bodies/{water_body_id}/indicators": {
   "get": {
    "tags": [
     "water-bodies"
    ],
    "summary": "Get Indicators",
    "description": "Indicator values for the latest scene on/before ``date``, with the\nDOY-matched baseline and the deviation, per zone.",
    "operationId": "get_indicators_api_v1_water_bodies__water_body_id__indicators_get",
    "parameters": [
     {
      "name": "water_body_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Water Body Id"
      }
     },
     {
      "name": "date",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "date"
        },
        {
         "type": "null"
        }
       ],
       "description": "latest scene on or before this date; default today",
       "title": "Date"
      },
      "description": "latest scene on or before this date; default today"
     },
     {
      "name": "zone",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Zone"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/IndicatorsResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/water-bodies/{water_body_id}/series": {
   "get": {
    "tags": [
     "water-bodies"
    ],
    "summary": "Get Series",
    "description": "Zone time series for charting, each point with its seasonal band and z.",
    "operationId": "get_series_api_v1_water_bodies__water_body_id__series_get",
    "parameters": [
     {
      "name": "water_body_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Water Body Id"
      }
     },
     {
      "name": "indicator",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "description": "one of ndti_turbidity, ndci_chlorophyll, fai_algal, sediment_proxy, mndwi_extent",
       "default": "ndti_turbidity",
       "title": "Indicator"
      },
      "description": "one of ndti_turbidity, ndci_chlorophyll, fai_algal, sediment_proxy, mndwi_extent"
     },
     {
      "name": "from",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "date"
        },
        {
         "type": "null"
        }
       ],
       "title": "From"
      }
     },
     {
      "name": "to",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "date"
        },
        {
         "type": "null"
        }
       ],
       "title": "To"
      }
     },
     {
      "name": "zone",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "defaults to the body's first zone",
       "title": "Zone"
      },
      "description": "defaults to the body's first zone"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SeriesResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "Feature": {
    "properties": {
     "type": {
      "type": "string",
      "const": "Feature",
      "title": "Type",
      "default": "Feature"
     },
     "id": {
      "type": "string",
      "title": "Id"
     },
     "geometry": {
      "additionalProperties": true,
      "type": "object",
      "title": "Geometry"
     },
     "properties": {
      "additionalProperties": true,
      "type": "object",
      "title": "Properties"
     }
    },
    "type": "object",
    "required": [
     "id",
     "geometry",
     "properties"
    ],
    "title": "Feature"
   },
   "FeatureCollection": {
    "properties": {
     "type": {
      "type": "string",
      "const": "FeatureCollection",
      "title": "Type",
      "default": "FeatureCollection"
     },
     "features": {
      "items": {
       "$ref": "#/components/schemas/Feature"
      },
      "type": "array",
      "title": "Features"
     }
    },
    "type": "object",
    "required": [
     "features"
    ],
    "title": "FeatureCollection"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "IndicatorsResponse": {
    "properties": {
     "water_body_id": {
      "type": "string",
      "title": "Water Body Id"
     },
     "requested_date": {
      "type": "string",
      "format": "date",
      "title": "Requested Date"
     },
     "scene_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Scene Id"
     },
     "observed_on": {
      "anyOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ],
      "title": "Observed On"
     },
     "zones": {
      "items": {
       "$ref": "#/components/schemas/ZoneIndicators"
      },
      "type": "array",
      "title": "Zones"
     },
     "cached": {
      "type": "boolean",
      "title": "Cached"
     },
     "disclaimer": {
      "type": "string",
      "title": "Disclaimer"
     }
    },
    "type": "object",
    "required": [
     "water_body_id",
     "requested_date",
     "scene_id",
     "observed_on",
     "zones",
     "cached",
     "disclaimer"
    ],
    "title": "IndicatorsResponse",
    "examples": [
     {
      "cached": false,
      "disclaimer": "Satellite-observed anomaly. Ground and laboratory testing recommended for validation.",
      "observed_on": "2026-09-17",
      "requested_date": "2026-09-17",
      "scene_id": "S2C_43QCA_20260917_0_L2A",
      "water_body_id": "wb_khadakwasla",
      "zones": [
       {
        "indicators": [
         {
          "baseline_mean": 0.128,
          "baseline_p10": 0.1,
          "baseline_p90": 0.16,
          "baseline_status": "usable",
          "baseline_std": 0.021,
          "deviation_pct": 143.8,
          "display_name": "Turbidity (NDTI)",
          "key": "ndti_turbidity",
          "p90": 0.41,
          "scientific_basis": "Red-to-green ratio; rises with sediment.",
          "valid_pixel_pct": 91.9,
          "value": 0.312,
          "water_fraction_pct": 97.0,
          "z_score": 8.76
         }
        ],
        "rejected": [],
        "zone_id": "wb_khadakwasla_z3",
        "zone_name": "Eastern zone"
       }
      ]
     }
    ]
   },
...
   "LatestObservation": {
    "properties": {
     "scene_id": {
      "type": "string",
      "title": "Scene Id"
     },
     "observed_on": {
      "type": "string",
      "format": "date",
      "title": "Observed On"
     },
     "cloud_pct": {
      "type": "number",
      "title": "Cloud Pct"
     },
     "usable": {
      "type": "boolean",
      "title": "Usable"
     },
     "stage": {
      "type": "string",
      "title": "Stage",
      "description": "deepest stage reached: ingested|masked|indicators|anomalies|scored|alerted"
     }
    },
    "type": "object",
    "required": [
     "scene_id",
     "observed_on",
     "cloud_pct",
     "usable",
     "stage"
    ],
    "title": "LatestObservation"
   },
   "ObservationItem": {
    "properties": {
     "scene_id": {
      "type": "string",
      "title": "Scene Id"
     },
     "observed_on": {
      "type": "string",
      "format": "date",
      "title": "Observed On"
     },
     "sensed_at": {
      "type": "string",
      "format": "date-time",
      "title": "Sensed At"
     },
     "platform": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Platform"
     },
     "cloud_pct": {
      "type": "number",
      "title": "Cloud Pct",
      "description": "tile-level cloud cover from STAC"
     },
     "usable": {
      "type": "boolean",
      "title": "Usable",
      "description": "mask verdict for this body (falls back to the tile flag)"
     },
     "valid_pixel_pct": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Valid Pixel Pct",
      "description": "cloud-free share over the body"
     },
     "water_extent_km2": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Water Extent Km2"
     },
     "stage": {
      "type": "string",
      "title": "Stage"
     },
     "source": {
      "type": "string",
      "title": "Source"
     }
    },
    "type": "object",
    "required": [
     "scene_id",
     "observed_on",
     "sensed_at",
     "platform",
     "cloud_pct",
     "usable",
     "valid_pixel_pct",
     "water_extent_km2",
     "stage",
     "source"
    ],
    "title": "ObservationItem",
    "examples": [
     {
      "cloud_pct": 8.1,
      "observed_on": "2026-09-17",
      "platform": "sentinel-2c",
      "scene_id": "S2C_43QCA_20260917_0_L2A",
      "sensed_at": "2026-09-17T05:31:12Z",
      "source": "earth-search",
      "stage": "alerted",
      "usable": true,
      "valid_pixel_pct": 91.9,
      "water_extent_km2": 17.2
     }
    ]
   },
   "ObservationList": {
    "properties": {
     "water_body_id": {
      "type": "string",
      "title": "Water Body Id"
     },
     "items": {
      "items": {
       "$ref": "#/components/schemas/ObservationItem"
      },
      "type": "array",
      "title": "Items"
     },
     "total": {
      "type": "integer",
      "title": "Total"
     },
     "next_cursor": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Next Cursor"
     }
    },
    "type": "object",
    "required": [
     "water_body_id",
     "items",
     "total",
     "next_cursor"
    ],
    "title": "ObservationList"
   },
   "SeriesPoint": {
    "properties": {
     "observed_at": {
      "type": "string",
      "format": "date-time",
      "title": "Observed At"
     },
     "scene_id": {
      "type": "string",
      "title": "Scene Id"
     },
     "value": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Value",
      "description": "zone mean"
     },
     "p90": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "P90"
     },
     "valid_pixel_pct": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Valid Pixel Pct"
     },
     "baseline_mean": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Baseline Mean"
     },
     "baseline_p10": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Baseline P10"
     },
     "baseline_p90": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Baseline P90"
     },
     "baseline_status": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Baseline Status"
     },
     "z_score": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Z Score"
     }
    },
    "type": "object",
    "required": [
     "observed_at",
     "scene_id",
     "value",
     "p90",
     "valid_pixel_pct",
     "baseline_mean",
     "baseline_p10",
     "baseline_p90",
     "baseline_status",
     "z_score"
    ],
    "title": "SeriesPoint"
   },
   "SeriesResponse": {
    "properties": {
     "water_body_id": {
      "type": "string",
      "title": "Water Body Id"
     },
     "zone_id": {
      "type": "string",
      "title": "Zone Id"
     },
     "indicator": {
      "type": "string",
      "title": "Indicator"
     },
     "display_name": {
      "type": "string",
      "title": "Display Name"
     },
     "date_from": {
      "type": "string",
      "format": "date",
      "title": "Date From"
     },
     "date_to": {
      "type": "string",
      "format": "date",
      "title": "Date To"
     },
     "points": {
      "items": {
       "$ref": "#/components/schemas/SeriesPoint"
      },
      "type": "array",
      "title": "Points"
     },
     "baseline_status": {
      "type": "string",
      "title": "Baseline Status"
     },
     "baseline_usable_windows": {
      "type": "integer",
      "title": "Baseline Usable Windows"
     },
     "weekly": {
      "items": {
       "additionalProperties": true,
       "type": "object"
      },
      "type": "array",
      "title": "Weekly",
      "description": "continuous-aggregate weekly means"
     },
     "cached": {
      "type": "boolean",
      "title": "Cached"
     },
     "disclaimer": {
      "type": "string",
      "title": "Disclaimer"
     }
    },
    "type": "object",
    "required": [
     "water_body_id",
     "zone_id",
     "indicator",
     "display_name",
     "date_from",
     "date_to",
     "points",
     "baseline_status",
     "baseline_usable_windows",
     "cached",
     "disclaimer"
    ],
    "title": "SeriesResponse",
    "examples": [
     {
      "baseline_status": "usable",
      "baseline_usable_windows": 366,
      "cached": false,
      "date_from": "2026-06-01",
      "date_to": "2026-09-17",
      "disclaimer": "Satellite-observed anomaly. Ground and laboratory testing recommended for validation.",
      "display_name": "Turbidity (NDTI)",
      "indicator": "ndti_turbidity",
      "points": [
       {
        "baseline_mean": 0.128,
        "baseline_p10": 0.1,
        "baseline_p90": 0.16,
        "baseline_status": "usable",
        "observed_at": "2026-09-17T05:31:12Z",
        "p90": 0.41,
        "scene_id": "S2C_43QCA_20260917_0_L2A",
        "valid_pixel_pct": 91.9,
        "value": 0.312,
        "z_score": 8.76
       }
      ],
      "water_body_id": "wb_khadakwasla",
      "weekly": [],
      "zone_id": "wb_khadakwasla_z3"
     }
    ]
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
   "WaterBodyDetail": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "district": {
      "type": "string",
      "title": "District"
     },
     "kind": {
      "type": "string",
      "title": "Kind"
     },
     "tier": {
      "type": "integer",
      "title": "Tier"
     },
     "area_km2": {
      "type": "number",
      "title": "Area Km2"
     },
     "centroid": {
      "items": {
       "type": "number"
      },
      "type": "array",
      "title": "Centroid",
      "description": "[lon, lat]"
     },
     "bbox": {
      "items": {
       "type": "number"
      },
      "type": "array",
      "title": "Bbox",
      "description": "[minlon, minlat, maxlon, maxlat]"
     },
     "n_zones": {
      "type": "integer",
      "title": "N Zones"
     },
     "latest_observation": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/LatestObservation"
       },
       {
        "type": "null"
       }
      ]
     },
     "status": {
      "type": "string",
      "enum": [
       "normal",
       "watch",
       "alert",
       "baseline_building",
       "no_data"
      ],
      "title": "Status"
     },
     "open_alerts": {
      "type": "integer",
      "title": "Open Alerts"
     },
     "max_open_severity": {
      "anyOf": [
       {
        "type": "string",
        "enum": [
         "low",
         "medium",
         "high"
        ]
       },
       {
        "type": "null"
       }
      ],
      "title": "Max Open Severity"
     },
     "baseline_status": {
      "type": "string",
      "enum": [
       "usable",
       "building",
       "none"
      ],
      "title": "Baseline Status"
     },
     "mgrs_tiles": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Mgrs Tiles"
     },
     "source": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Source"
     },
     "boundary": {
      "additionalProperties": true,
      "type": "object",
      "title": "Boundary",
      "description": "GeoJSON MultiPolygon, EPSG:4326"
     },
     "zones": {
      "$ref": "#/components/schemas/FeatureCollection"
     },
     "disclaimer": {
      "type": "string",
      "title": "Disclaimer"
     }
    },
    "type": "object",
    "required": [
     "id",
     "name",
     "district",
     "kind",
     "tier",
     "area_km2",
     "centroid",
     "bbox",
     "n_zones",
     "latest_observation",
     "status",
     "open_alerts",
     "max_open_severity",
     "baseline_status",
     "mgrs_tiles",
     "source",
     "boundary",
     "zones",
     "disclaimer"
    ],
    "title": "WaterBodyDetail",
    "examples": [
     {
      "area_km2": 18.4,
      "baseline_status": "usable",
      "bbox": [
       73.72,
       18.4,
       73.82,
       18.47
      ],
      "centroid": [
       73.7712,
       18.4419
      ],
      "district": "Pune",
      "id": "wb_khadakwasla",
      "kind": "reservoir",
      "latest_observation": {
       "cloud_pct": 8.1,
       "observed_on": "2026-09-17",
       "scene_id": "S2C_43QCA_20260917_0_L2A",
       "stage": "alerted",
       "usable": true
      },
      "max_open_severity": "high",
      "n_zones": 6,
      "name": "Khadakwasla Reservoir",
      "open_alerts": 1,
      "status": "alert",
      "tier": 1
     }
    ]
   },
   "WaterBodyList": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/WaterBodyListItem"
      },
      "type": "array",
      "title": "Items"
     },
     "total": {
      "type": "integer",
      "title": "Total"
     },
     "next_cursor": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Next Cursor"
     }
    },
    "type": "object",
    "required": [
     "items",
     "total",
     "next_cursor"
    ],
    "title": "WaterBodyList"
   },
   "WaterBodyListItem": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "district": {
      "type": "string",
      "title": "District"
     },
     "kind": {
      "type": "string",
      "title": "Kind"
     },
     "tier": {
      "type": "integer",
      "title": "Tier"
     },
     "area_km2": {
      "type": "number",
      "title": "Area Km2"
     },
     "centroid": {
      "items": {
       "type": "number"
      },
      "type": "array",
      "title": "Centroid",
      "description": "[lon, lat]"
     },
     "bbox": {
      "items": {
       "type": "number"
      },
      "type": "array",
      "title": "Bbox",
      "description": "[minlon, minlat, maxlon, maxlat]"
     },
     "n_zones": {
      "type": "integer",
      "title": "N Zones"
     },
     "latest_observation": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/LatestObservation"
       },
       {
        "type": "null"
       }
      ]
     },
     "status": {
      "type": "string",
      "enum": [
       "normal",
       "watch",
       "alert",
       "baseline_building",
       "no_data"
      ],
      "title": "Status"
     },
     "open_alerts": {
      "type": "integer",
      "title": "Open Alerts"
     },
     "max_open_severity": {
      "anyOf": [
       {
        "type": "string",
        "enum": [
         "low",
         "medium",
         "high"
        ]
       },
       {
        "type": "null"
       }
      ],
      "title": "Max Open Severity"
     },
     "baseline_status": {
      "type": "string",
      "enum": [
       "usable",
       "building",
       "none"
      ],
      "title": "Baseline Status"
     }
    },
    "type": "object",
    "required": [
     "id",
     "name",
     "district",
     "kind",
     "tier",
     "area_km2",
     "centroid",
     "bbox",
     "n_zones",
     "latest_observation",
     "status",
     "open_alerts",
     "max_open_severity",
     "baseline_status"
    ],
    "title": "WaterBodyListItem",
    "examples": [
     {
      "area_km2": 18.4,
      "baseline_status": "usable",
      "bbox": [
       73.72,
       18.4,
       73.82,
       18.47
      ],
      "centroid": [
       73.7712,
       18.4419
      ],
      "district": "Pune",
      "id": "wb_khadakwasla",
      "kind": "reservoir",
      "latest_observation": {
       "cloud_pct": 8.1,
       "observed_on": "2026-09-17",
       "scene_id": "S2C_43QCA_20260917_0_L2A",
       "stage": "alerted",
       "usable": true
      },
      "max_open_severity": "high",
      "n_zones": 6,
      "name": "Khadakwasla Reservoir",
      "open_alerts": 1,
      "status": "alert",
      "tier": 1
     }
    ]
   },
...
   "ZoneIndicators": {
    "properties": {
     "zone_id": {
      "type": "string",
      "title": "Zone Id"
     },
     "zone_name": {
      "type": "string",
      "title": "Zone Name"
     },
     "indicators": {
      "items": {
       "$ref": "#/components/schemas/ZoneIndicator"
      },
      "type": "array",
      "title": "Indicators"
     },
     "rejected": {
      "items": {
       "additionalProperties": true,
       "type": "object"
      },
      "type": "array",
      "title": "Rejected",
      "description": "zone-indicator records L6 refused, with reason"
     }
    },
    "type": "object",
    "required": [
     "zone_id",
     "zone_name",
     "indicators"
    ],
    "title": "ZoneIndicators"
   },
...
  }
 }
}